% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{display_quiz_question}
\alias{display_quiz_question}
\title{Display quiz question}
\usage{
display_quiz_question(
  place_name,
  question_text,
  correct_answer,
  incorrect_choice_1,
  incorrect_choice_2,
  incorrect_choice_3
)
}
\arguments{
\item{place_name}{String of the haunted place name}

\item{question_text}{String of the question text}

\item{correct_answer}{String with the correct answer to the quiz question}

\item{incorrect_choice_1}{String with the first incorrect answer for the quic}

\item{incorrect_choice_2}{String with the second incorrect answer for the quic}

\item{incorrect_choice_3}{String with the third incorrect answer for the quic}
}
\value{
a list with the following elements:
* place_name
* question_text
* correct_answer
* incorrect_choice_1
* incorrect_choice_2
* incorrect_choice_3
}
\description{
This function assembles the necessary information to display a quiz question and multiple choice answers
}
