% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_user_prompt}
\alias{prepare_user_prompt}
\title{Prepare user prompt}
\usage{
prepare_user_prompt(location, latitude, longitude, state, description)
}
\arguments{
\item{location}{name of haunted place}

\item{latitude}{haunted place latitude number}

\item{longitude}{haunted place longitude number}

\item{state}{string of state (full name) where haunted place is located}

\item{description}{string with description of haunted place}
}
\value{
string with user prompt to send to chat bot
}
\description{
Prepare user prompt
}
