% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_haunted_data}
\alias{process_haunted_data}
\title{process haunted data frame}
\usage{
process_haunted_data()
}
\arguments{
\item{n_records}{Number of records to retain for each state: Default is 4}
}
\value{
duckplyr tibble of process haunted places data frame
}
\description{
process haunted data frame
}
